+++ test/schema_parser_test.py
@@ -26,6 +26,7 @@ from brewtils.test.comparable import (
     assert_parameter_equal,
     assert_patch_equal,
     assert_queue_equal,
+    assert_replication_equal,
     assert_request_equal,
     assert_request_file_equal,
     assert_resolvable_equal,
@@ -216,6 +217,12 @@ class TestParse(object):
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
             ),
+            (
+                brewtils.models.Replication,
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_single(self, model, data, assertion, expected):
@@ -380,6 +387,12 @@ class TestParse(object):
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
             ),
+            (
+                "parse_replication",
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_single_specific(self, method, data, assertion, expected):
@@ -537,6 +550,12 @@ class TestParse(object):
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
             ),
+            (
+                brewtils.models.Replication,
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_many(self, model, data, assertion, expected):
@@ -687,6 +706,12 @@ class TestParse(object):
                 assert_topic_equal,
                 lazy_fixture("bg_topic"),
             ),
+            (
+                "parse_replication",
+                lazy_fixture("replication_dict"),
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_many_specific(self, method, data, assertion, expected):
@@ -745,6 +770,7 @@ class TestSerialize(object):
             (lazy_fixture("bg_resolvable"), lazy_fixture("resolvable_dict")),
             (lazy_fixture("bg_subscriber"), lazy_fixture("subscriber_dict")),
             (lazy_fixture("bg_topic"), lazy_fixture("topic_dict")),
+            (lazy_fixture("bg_replication"), lazy_fixture("replication_dict")),
         ],
     )
     def test_single(self, model, expected):
@@ -871,6 +897,11 @@ class TestSerialize(object):
                 lazy_fixture("bg_topic"),
                 lazy_fixture("topic_dict"),
             ),
+            (
+                "serialize_replication",
+                lazy_fixture("bg_replication"),
+                lazy_fixture("replication_dict"),
+            ),
         ],
     )
     def test_single_specific(self, method, data, expected):
@@ -904,6 +935,7 @@ class TestSerialize(object):
             (lazy_fixture("bg_resolvable"), lazy_fixture("resolvable_dict")),
             (lazy_fixture("bg_subscriber"), lazy_fixture("subscriber_dict")),
             (lazy_fixture("bg_topic"), lazy_fixture("topic_dict")),
+            (lazy_fixture("bg_replication"), lazy_fixture("replication_dict")),
         ],
     )
     def test_many(self, model, expected):
@@ -1013,6 +1045,11 @@ class TestRoundTrip(object):
                 lazy_fixture("bg_subscriber"),
             ),
             (brewtils.models.Topic, assert_topic_equal, lazy_fixture("bg_topic")),
+            (
+                brewtils.models.Replication,
+                assert_replication_equal,
+                lazy_fixture("bg_replication"),
+            ),
         ],
     )
     def test_parsed_start(self, model, assertion, data):
@@ -1046,6 +1083,7 @@ class TestRoundTrip(object):
             (brewtils.models.Operation, lazy_fixture("operation_dict")),
             (brewtils.models.Runner, lazy_fixture("runner_dict")),
             (brewtils.models.Resolvable, lazy_fixture("resolvable_dict")),
+            (brewtils.models.Replication, lazy_fixture("replication_dict")),
         ],
     )
     def test_serialized_start(self, model, data):